Refactor reviews.lua
- refactor thread buffer actions to threat-panel.lua
- refactor some reviews related functions into the Review class as member
methods
