+++ plugin/octo.vim
@@ -12,8 +12,8 @@ augroup octo_autocmds
   au BufReadCmd octo://* lua require'octo'.load_buffer()
   au BufWriteCmd octo://* lua require'octo'.save_buffer()
   au CursorHold octo://* lua require'octo'.on_cursor_hold()
-  au CursorHold * lua require'octo.reviews'.show_review_threads()
-  au CursorMoved * lua require'octo.reviews'.hide_review_threads()
+  au CursorHold * lua require'octo.reviews.thread-panel'.show_review_threads()
+  au CursorMoved * lua require'octo.reviews.thread-panel'.hide_review_threads()
   au TabClosed * lua require'octo.reviews'.close(tonumber(vim.fn.expand("<afile>")))
   au TabLeave * lua require'octo.reviews'.on_tab_leave()
   au WinLeave * lua require'octo.reviews'.on_win_leave()
